Address PR review: Consolidate component files per project style
- Rename index.tsx to llmbot_post.tsx (avoid index pattern)
- Create permalink_data.ts with PermalinkData type and utility
- Move types inline to files that own them (no separate types.ts)
- Move styled components inline to files that own them (no separate styles.tsx)
- Update all imports to reference new file structure
