+++ webapp/src/components/llmbot_post/permalink_data.ts
@@ -1,7 +1,16 @@
 // Copyright (c) 2023-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {PermalinkData} from './types';
+export type PermalinkData = {
+    channel_display_name: string
+    channel_id: string
+    post_id: string
+    team_name: string
+    post: {
+        message: string
+        user_id: string
+    }
+}
 
 export function extractPermalinkData(post: any): PermalinkData | null {
     for (const embed of post?.metadata?.embeds || []) {
